backend/requirements.txt
fastapi
uvicorn
requests
sqlalchemy
python-dotenv
pydantic
